* moves css files into components dir
