+++ packages/src/components/button/button.module.scss
@@ -1,4 +1,4 @@
-@use '../mixin' as *;
+@use '../common/mixin' as *;
 
 .button {
   border: 1px solid var(--uil-black-light);
