+++ packages/src/components/check/check.module.scss
@@ -1,4 +1,4 @@
-@use '../mixin' as *;
+@use '../common/mixin' as *;
 
 .checkWrapper {
   display: flex;
