+++ packages/src/components/check/checkbox/checkbox.module.scss
@@ -1,4 +1,4 @@
-@use '../../mixin' as *;
+@use '../../common/mixin' as *;
 
 .checkbox {
   border-radius: 2px;
