+++ packages/src/components/check/radio/radio.module.scss
@@ -1,4 +1,4 @@
-@use '../../mixin' as *;
+@use '../../common/mixin' as *;
 
 .radioWrapper {
   display: flex;
