+++ packages/src/components/common/variables.scss
