+++ packages/src/components/input/customInput.tsx
@@ -3,7 +3,7 @@ import {BaseInput, type IBaseInput} from './baseInput';
 import {Icon} from 'components/icon/icon';
 import {useClickOutsideRef} from 'hooks/clickOutside';
 import style from './input.module.scss';
-import global from '../global.module.scss';
+import global from '../common/global.module.scss';
 
 export interface ICustomInput extends IBaseInput {
   tooltipClose?: string;
