+++ packages/src/components/input/input.module.scss
@@ -1,4 +1,4 @@
-@use '../mixin' as *;
+@use '../common/mixin' as *;
 
 .inputWrapper {
   position: relative;
