+++ packages/src/components/modal/modal.module.scss
@@ -1,4 +1,4 @@
-@use '../mixin' as *;
+@use '../common/mixin' as *;
 
 .modalWrapper {
   position: fixed;
