--- packages/src/components/variables.scss.d.ts
@@ -1,2 +1,0 @@
-declare const cssString: string;
-export default cssString;
